refactor: legacy controller와 어노테이션 기반 controller 분리
