+++ app/src/main/java/com/techcourse/controller/legacy/LoginController.java
@@ -1,4 +1,4 @@
-package com.techcourse.controller;
+package com.techcourse.controller.legacy;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
